requirements.txt
# Django Core
Django>=4.2,<5.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0

# Database
psycopg2-binary>=2.9.9

# Authentication & Security
djangorestframework-simplejwt>=5.3.0
django-auditlog>=2.3.0

# PDF Generation
reportlab>=4.0.0
WeasyPrint>=60.0
PyPDF2>=3.0.0
Pillow>=10.0.0

# Barcode Generation
python-barcode>=0.15.1
qrcode>=7.4.2

# Excel Export
openpyxl>=3.1.2

# Email
django-anymail>=10.2

# Task Queue (for background jobs like backup)
celery>=5.3.0
redis>=5.0.0

# Utilities
python-dateutil>=2.8.2
python-dotenv>=1.0.0

# Development
django-debug-toolbar>=4.2.0

# Production
gunicorn>=21.2.0
whitenoise>=6.6.0
